Colored & titled envelopes. Arrow detections. No Axes/Figure aliases.

+++ sharp/data/files/figure.py
@@ -1,5 +1,7 @@
+from matplotlib.figure import Figure
+
 from sharp.data.files.base import OutputFileTarget
-from sharp.data.types.aliases import Figure
+from sharp.data.files.config import data_config
 
 
 class MatplotlibFigureFile(OutputFileTarget):
@@ -21,15 +23,20 @@ class FigureTarget(OutputFileTarget):
     """
 
     def exists(self):
-        return self.bitmap_version.exists() and self.vector_version.exists()
+        if data_config.bitmap_versions:
+            return self.vector_version.exists() and self.bitmap_version.exists()
+        else:
+            return self.vector_version.exists()
 
     def write(self, fig: Figure):
-        self.bitmap_version.write(fig)
         self.vector_version.write(fig)
+        if data_config.bitmap_versions:
+            self.bitmap_version.write(fig)
 
     def delete(self):
-        self.bitmap_version.delete()
         self.vector_version.delete()
+        if data_config.bitmap_versions:
+            self.bitmap_version.delete()
 
     @property
     def bitmap_version(self):
